backend/cbt_sessions/serializers.py
from rest_framework import serializers
from .models import ExamSession, Answer

class ExamSessionSerializer(serializers.ModelSerializer):
    class Meta:
        model = ExamSession
        fields = '__all__'

class AnswerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Answer
        fields = '__all__'